Khagastara's branch

+++ database/migrations/2025_06_05_064625_create_admins_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('admin_name');
             $table->string('phone_number')->unique();
-            $table->unsignedInteger('id_account');
+            $table->unsignedBigInteger('id_account');
             $table->timestamps();
 
             $table->foreign('id_account')->on('accounts')->references('id');
